@@ -1,14 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
+        <LangVersion>8</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Lamar" Version="3.2.0" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.2.0" />
-      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.3.0" />
-      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.3.0" />
+      <PackageReference Include="Lamar" Version="4.1.0" />
+      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.1.2" />
+      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.6.0" />
+      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.6.0" />
     </ItemGroup>
 
 </Project>
