@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.AspNetCore.Mvc.Versioning;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.Filters;
 using Swashbuckle.AspNetCore.Swagger;
 using Swashbuckle.AspNetCore.SwaggerGen;
@@ -43,25 +44,26 @@ namespace plodo.Backend.API.Configurations
                 
             options.OperationFilter<SwaggerDefaultValues>();
 
-            options.AddSecurityDefinition("oauth2", new ApiKeyScheme()
+            options.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
             {
-                Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
+                Description =
+                    "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
                 Name = "Authorization",
-                In = "header",
-                Type = "apiKey"
+                In = ParameterLocation.Header,
+                Type = SecuritySchemeType.ApiKey,
+                Scheme = "Bearer"
             });
         }
         
-        static Info CreateInfoForApiVersion( ApiVersionDescription description )
+        static OpenApiInfo CreateInfoForApiVersion( ApiVersionDescription description )
         {
-            var info = new Info()
+            var info = new OpenApiInfo()
             {
                 Title = "plodo API",
                 Version = description.ApiVersion.ToString(),
                 Description = "Backend API for the plodo.io realtime audience feedback",
-                Contact = new Contact() { Name = "Lars Baunwall", Email = "larslb@plodo.io" },
-                TermsOfService = "Commercial",
-                License = new License() { Name = "Commercial", Url = "" }
+                Contact = new OpenApiContact() { Name = "Lars Baunwall", Email = "larslb@plodo.io" },
+                License = new OpenApiLicense() { Name = "Commercial" }
             };
 
             if ( description.IsDeprecated )
@@ -84,7 +86,7 @@ namespace plodo.Backend.API.Configurations
     }
     public class SwaggerDefaultValues : IOperationFilter
     {
-        public void Apply( Operation operation, OperationFilterContext context )
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             var apiDescription = context.ApiDescription;
             var apiVersion = apiDescription.GetApiVersion();
@@ -97,22 +99,28 @@ namespace plodo.Backend.API.Configurations
                 return;
             }
 
-            foreach ( var parameter in operation.Parameters.OfType<NonBodyParameter>() )
+            foreach ( var parameter in operation.Parameters )
             {
                 var description = apiDescription.ParameterDescriptions
                     .First( p => p.Name == parameter.Name );
-            
+                var routeInfo = description.RouteInfo;
+                
+                if (string.IsNullOrEmpty(parameter.Name))
+                {
+                    parameter.Name = description.ModelMetadata?.Name;
+                }
+                
                 if ( parameter.Description == null )
                 {
                     parameter.Description = description.ModelMetadata?.Description;
                 }
 
-                if ( parameter.Default == null )
+                if (routeInfo == null)
                 {
-                    parameter.Default = description.DefaultValue;
+                    continue;
                 }
 
-                parameter.Required |= description.IsRequired;
+                parameter.Required |= !routeInfo.IsOptional;
             }
         }
     }
